abstract waitlist form for other use
